+++ sdk/src/model/CreateSelectObjectMetaResult.cc
@@ -20,7 +20,7 @@
 #include "../utils/Crc32.h"
 
 #define FRAME_HEADER_LEN   (12+8)
-#define PARSE_FOUR_BYTES(a, b, c, d) (((uint64_t)(a) << 24)|((uint64_t)(b) << 16)|((uint64_t)(c) << 8)|(d))
+#define PARSE_FOUR_BYTES(a, b, c, d) (((uint32_t)(a) << 24)|((uint32_t)(b) << 16)|((uint32_t)(c) << 8)|(d))
 #define PARSE_EIGHT_BYTES(a, b, c, d, e, f, g, h) (((uint64_t)(a)<<56)|((uint64_t)(b)<<48)| ((uint64_t)(c)<<40)| ((uint64_t)(d)<<32)| ((uint64_t)(e)<<24)| ((uint64_t)(f)<<16)| ((uint64_t)(g)<<8)| (h))
 
 using namespace AlibabaCloud::OSS;
@@ -62,7 +62,7 @@ CreateSelectObjectMetaResult& CreateSelectObjectMetaResult::operator=(const std:
         int type = 0;
         type = (buffer[1] << 16) | (buffer[2] << 8) | buffer[3];
         // payload length 4 bytes
-        int length = PARSE_FOUR_BYTES(buffer[4], buffer[5], buffer[6], buffer[7]);
+        uint32_t length = PARSE_FOUR_BYTES(buffer[4], buffer[5], buffer[6], buffer[7]);
         // header checksum
 
         // payload
@@ -72,7 +72,7 @@ CreateSelectObjectMetaResult& CreateSelectObjectMetaResult::operator=(const std:
         case 0x800007:
         {
             uint32_t payloadCrc32 = 0;
-            int messageLength = length - 32;
+            uint32_t messageLength = length - 32;
             data->read(reinterpret_cast<char*>(buffer), 32);
             payloadCrc32 = CRC32::CalcCRC(payloadCrc32, buffer, 32);
             // offset 8 bytes
